[Concurrency] Emit extended method type descriptors for @objc async methods.
For `@objc async` methods, all of the important information about the
"result" type and whether it throws or not is within the type of the
completion handler block. Enable extended type information for the
generated Objective-C metadata of such methods, so that clients can
understand how to invoke them using reflection facilities.

Implements rdar://73048313.