[AST Verifier] Handle ObjC requirements satisfied by alternative requirements.
With 'async' imports of Objective-C protocol requirements, a single
Objective-C method will produce two different requirements in the
protocol, with the same Objective-C selector. Teach the AST verifier
that it is sufficient for just one of the requirements with that
Objective-C selector to be satisfied by a witness.

Fixes rdar://72742910.