[test] Allow setting extra environment for running tests.
In some situations, in order to set the correct environment during some
tests, extra environment has to be provided to the LLVM Lit invocations.
It seems that the preferred way is using `cmake -E env` to be
cross-platform.

The environment can be set using a cached variable so it can be set from
CMake cache files, or passing parameters to the CMake configuration
step.