Add test case for `async try foo()`
Ensure that we emit the proper diagnostics for when the await and try
are flipped, but also await is incorrectly spelled `async`.

We're expecting two diagnostics, both of which have fix-its that do the
right thing. One will just replace the `async` with an `await`. The
other will move the text so that try comes first. Because `await` and
`async` are both have the same number of characters, both fix-its will
fix the issue correctly.