[sil] Disable MemoryLifetimeVerifier on alloc_stacks that are enum typed.
Currently in this verifier, we stop verifying if we find a switch_enum_addr
use. This creates a problem since no one has gone through and changed the
frontend/optimizer to understand that it needs to insert destroy_addr on
alloc_stack even if dynamically we know that the enum has a trivial case or
non-payloaded case due to a switch_enum_addr. So if one then performs the
completely unrelated, valid optimization of promoting the switch_enum_addr to a
switch_enum (lets say using a load_borrow), then this verifier will see a leaked
value along the non-payloaded path.

Disable this verification for now until a complete analysis of enums is provided
to unblock further work.

https://bugs.swift.org/browse/SR-14123