Try to fix the ASAN build
We want to use the 'just built' compiler for the runtime tests because
of the recently added `swiftasync` parameter support.

But ASAN won't link if we use the 'just built' compiler because libgtest
and LLVMSupport are compiled by the host compiler.

Disable the runtime test when running under ASAN.

rdar://73664504