[Build] Fix macOS version passed in -target parameter.
This code was added in commit b4368259480ed2c895e8c855bbfde20c1805acd1, but the macOS variable expansion had an extra "SWIFT_" at the front.

Among other potential problems, this causes classes in the stdlib to always be built with the legacy is-swift bit, which causes havoc when libswiftCore is in the shared cache where those bits aren't fixed up.

rdar://73767816