[IRGen] Return unsigned context from resume_project_context.
In __swift_async_resume_project_context, the context is stored into the
extended frame.  On arm64e, it is signed first.  Previously, that signed
context was returned fro the function.  That resulted in code like

    pacda  x16, x10
    str    x16, [x9]
    ldr    x9, [x16, #0x48]

where the context is signed (pacda), stored into the extended frame
(str) and then an attempt is made to load from the signed context (ldr).
Here, the unsigned context is returned from the function.