[5.4] Fix OSLogOptimization pass to mark array.finalize_intrinsic as not foldable
(This is a cherry-pick of PR #35636 to the release/5.4 branch.)

This PR fixes a compiler crash caused due to an assertion failure
in the  OSLogOptimization pass that happens when optimizing the
`os_signpost(.animationBegin, ...)`  API introduced in iOS 14
for measuring performance of animations.  A combination of recent
changes to the compiler in Swift 5.4, including introduction
of a new  _finalizeUninitializedArray intrinsic, caused the
optimization pass to analyze more code than necessary for that API,
which caused an invariant assumed by the pass to fail. This change
relaxes the assumption and makes the pass take an early exit when
the invariants do not hold, which prevents it from crashing in
such cases.

Resolves rdar://72634574