[sourcekitd-test] Add option to count the number of instructions executed during a request
The number of instructions executed gives a pretty stable measurement
that's a good proxy for the time an operation took.

To use this in sourcekitd-test, we need to use an in-process sourcekitd
(since sourcekitd-test cannot by itself inspect the instruction counter)
of an XPC service.

I'm putting this behind an additional flag for now because whoever uses
it, should do so deliberately and understand the limitation that this
counter only gives meaningful results when used together with an
in-process sourcekitd.