flush delayed diagnostics in dtor of MultiConformanceChecker
While individual ConformanceCheckers tend to be used for one-off
conformance queries where diagnostics are fully suppressed, the
MultiConformanceChecker is a major driver of overall protocol
conformance checking and should not be losing delayed error diags.

In some situations (mostly with checking for conformance to ObjC
protocols that have async-like requirements), a delayed error
diagnostic can be lost by the MultiConformanceChecker, because
the deferred diag was added after the calls to emit them happen.
This new flush operation in MultiConformanceChecker's destructor
fixes the issue in those situations.

Additionally, I attempted to try and catch mistakes like this in
the future by adding an assert to ConformanceChecker's dtor for
obviously lost error diagnostics.

resolves rdar://73641790