[ownership] Change the stdlib to serialize code in ossa form on macOS.
I am doing it this way so that I can:

1. There is some sort of ASAN issue that this exposes on Linux, so I am going to do
this on Darwin and then debug the Linux issue using ASAN over the weekend/next
week.

2. We are seeing I think large CFGs on 32-bit platforms since we are not
simplifying some sort of pattern. I don't have time to figure that out right
now, so I am disabling it on non-macOS platforms so we can get something in.