Sema: Reject non-stored local variables in result builders
We would previously reject properties without an initializer expression,
but this is not quite sufficient, because of 'lazy', property wrappers,
and observers.

We could probably support all three in result builder contexts, but
for now, let's diagnose them like other computed properties, instead of
crashing.

Fixes part of <rdar://problem/73545981>.