[partial-spec] Turn off partial generic specialization on the stdlib/foundation.
The reason why I am doing this is that we only have it enabled on the
stdlib/foundation. This causes it to not have the same test coverage that other
forms of inlining since those forms of inlining run anywhere.

I recently just hit a bug due to this so I was curious how much bang we are
getting for this.