Fix copy-propagation: bail-out on reborrows.
The simple logic for discovering pruned liveness and skipping over
nested borrow scopes assumed a dominating definition. This isn't the
case for arbitrary reborrows and would potentially lead to OSSA
verification errors.

Rather than add support to handle certain cases, bail-out on this
situation, which is rare enough that it isn't a priority to optimize
yet.

Noticed by inspection while working on the OSSA utilities.