SIL: Pick default ARC convention for enum case witness thunks based on the kind of the original decl.
Enum constructors take their arguments +1 by default, but they can now be used to satisfy protocol static member
requirements, which take arguments +0 by default. Type lowering would accidentally use the kind of the witness
to determine the conventions for the witness thunk, leading to a miscompile when the requirement is called through
the protocol. This differs from the fix on main in that the behavior change is contained to only cases where the
witness is an enum constructor, to minimize behavior change. rdar://73855052